+++ NEWS.md
@@ -1,3 +1,9 @@
+# kv 0.2.0 
+
+* Added consistent behavior for missing NAMES
+* Added `test/testthat/test-missing-names.R`
+* Updated man
+
 # kv 0.1.0
 
 * Added a `NEWS.md` file to track changes to the package.